simplified project a little more
